+++ src/xarray_kat/katdal_types.py
@@ -1,16 +1,26 @@
-from katsdptelstate import TelescopeState
+from __future__ import annotations
 
-from xarray_kat.multiton import Multiton
-from xarray_kat.third_party.vendored.katdal.applycal_minimal import CorrectionParams
+from typing import TYPE_CHECKING, List
+
+from katpoint import Antenna
+
 from xarray_kat.third_party.vendored.katdal.datasources_minimal import (
   TelstateDataSource,
 )
-from xarray_kat.third_party.vendored.katdal.sensordata import SensorCache
+
+# Make these symbols available
 from xarray_kat.third_party.vendored.katdal.vis_flags_weights_minimal import (
   AutoCorrelationIndices,  # noqa: F401
   corrprod_to_autocorr,  # noqa: F401
 )
 from xarray_kat.third_party.vendored.katdal.visdatav4_minimal import VisibilityDataV4
+
+if TYPE_CHECKING:
+  from katsdptelstate import TelescopeState
+
+  from xarray_kat.multiton import Multiton
+  from xarray_kat.third_party.vendored.katdal.applycal_minimal import CorrectionParams
+  from xarray_kat.third_party.vendored.katdal.sensordata import SensorCache
 
 
 class TelstateDataProducts:
@@ -33,6 +43,21 @@ class TelstateDataProducts:
   def telstate(self) -> TelescopeState:
     """Return the TelescopeState"""
     return self.datasource.telstate
+
+  @property
+  def antennas(self) -> List[Antenna]:
+    """Return a list of katpoint Antenna"""
+    antennas = []
+    ts = self.telstate
+
+    for resource in ts["sub_pool_resources"].split(","):
+      try:
+        antenna_description = ts[f"{resource}_observer"]
+        antennas.append(Antenna(antenna_description))
+      except (KeyError, ValueError):
+        continue
+
+    return antennas
 
   @property
   def sensor_cache(self) -> SensorCache: